some minor refactoring
